TASK: Move union-type tests into main suite as we only support php 8 nowadays
